extract functions outside of the main file
