+++ q2q.js
@@ -1,32 +1,3 @@
-#! /usr/bin/env node
-
-const fs   = require("fs");
-const args = process.argv;
-
-// Check arguments
-if (!(args[2] && args[3] && args[4]) || (args[2] !== "single" && args[2] !== "double")) {
-  console.log(`
-Usage: q2q [single|double] src.js out.js
-`);
-  
-  process.exit();
-}
-
-// Load the JS file transform it
-const pref  = args[2] === "single" ? "'" : '"';
-const js    = fs.readFileSync(args[3], {encoding: "utf8"});
-const outJS = switchStringLiterals(js, pref);
-
-// Save to dest
-fs.writeFileSync(args[4], outJS, {encoding: "utf8"});
-
-/*
-  
-  This is Quote-to-Quote's source,
-  this is a comment to show it ignores comments
-  
- */
-
 function switchStringLiterals(rawSource, rawPreference) {
   const source = rawSource;
   const prefer = rawPreference === "'" || rawPreference === '"' ? rawPreference : '"';
@@ -184,3 +155,6 @@ function opposite(quote) {
     return '"';
   }
 }
+
+module.exports.switchStringLiterals = switchStringLiterals;
+module.exports.opposite = opposite;
